# Conflicts:
#	src/main/__main__.py
#	src/parser/ASTConversion.py

+++ TestCases/LLVMTests/FunctionTests/tests/test5.c
@@ -14,7 +14,7 @@ int main() {
 
 
 
-int func(int a, int b) {
+int func(float a, int b) {
     return a + b;
 }
